AID Persistence, Add/Remove Accessories, Reachability (#40)
Add ability to add and remove accessories from the Bridge (while paired or not), and persist the Accessory IDs over restarts of the host application. The TXT record broadcast over mDNS is updated when the configuation changes. Add a reachable flag to Accessory, for subclasses to indicate that the bridge has lost contact with a device, which will be reflected in the status results sent to HomeKit. If the Accessory has a BridgingState Service, then the reachability status is updated too. pairing/unpairing is now performed through functions on Device, so the status can be monitored.