Use stable hash to determine if configuration number needs bumping
The old implementation would increase the configuration number on each
server restart. This caused issues with e.g. Homepod that would reset
room configurations. We need a stable hash that is the same accross reboots
and maybe even hardware. We don't need a good distribution of this hash, but
importantly a hash that is unique for changes to device configuration.